+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## 0.4.33
 
+* `Versions`:
+    * `Ktor`: `1.5.2` -> `1.5.3`
 * `Coroutines`
     * Add `WeakJob` workaround:
         * `CoroutineScope#weakLaunch`